完成
ほぼ完成だと思われる。

+++ ge_p6/src/pop.c
@@ -7,7 +7,7 @@
 #include <stdio.h>
 #include "stack.h"
 
-int pop(struct stack *five){
+long long int pop(struct stack *five){
     five->head--;//スタックの先頭のインデックスを変更
     printf("スタック[%d]の%lldをpop\n",five->head,five->value[five->head]);//スタック先頭の数値を格納
     return five->value[five->head];//スタック先頭の数値を格納
